Rework frame event handling in ConditionAction class
This commit simplifies the event binding for frame enter and leave
actions by directly binding the events to the frame, removing
unnecessary variables and unbinding logic.

On my machine flickering was no issue anymore as soon the deactivated
frame is not bigger than the activated frame. I tested this on my
machine.