Fix the hover offset
With this change the inner position of the labels and edits doesn't move
around when hovered by a mouse.

The adjustment of the borderwidth changes the position of the inner
conent. This leads to a weird jumpy behavior. With this change, we
adjust the window position slightly, so that the inner content stays at
the same position on a mouse hover. (No more jumping around).

* state actions
* comments
* condition_action_handling