Introduce generation path query functions and use them
This refactors the path representation inside the application.
Previously, str was used to represent a path. Now with this change the
hdl generation and many parts that require a path now use the
pathlib.Path class instead.

This has the following benefits: * The typesystem tells clearly that
something is a Path * There are functions to convert a path to a posix
path * Path comparison works out of the box (on Windows '\' is the same
as '/')